- made `size` prop optional since we have default value for this prop - rewritten `CheckboxDropdownTest` in typescript - refactored/optimized `CheckboxDropdownTest.tsx`
